pyproject.toml
[project]
name = "ggs"
version = "0.1.0"
description = "Guru Granth Sahib — Structured Multi-Layer Textual Analysis Platform"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "Harpreet Singh" }]

dependencies = [
    "pyyaml>=6.0",
    "rich>=14.0",
    "typer>=0.15",
    "pyahocorasick>=2.3",
    "beautifulsoup4>=4.13",
    "lxml>=5.3",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "hypothesis>=6.120",
    "ruff>=0.11",
    "coverage>=7.6",
]

[project.scripts]
ggs = "ggs.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ggs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"

[tool.ruff]
target-version = "py314"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
# RUF001/RUF003: Gurmukhi Unicode is the primary content of this project,
# not "ambiguous" characters.  These rules are designed for Latin-script
# projects where Cyrillic/Greek lookalikes sneak in accidentally.
ignore = ["RUF001", "RUF002", "RUF003"]
